frontend for api/v1/invites#create
